[OB-3917] fix!: change all int_64t exposed script parameters to int_32t (#624)
This is because our current version of quickjs does not support BigInt
(64 bits ints), which led to an error if using BigInt in a script on
Web, or an unsupported conversion warning if using a Number.

Now that we expose 32 bits ints, the web scripts will no longer show an
unsupported conversion warning when using a Number.

BREAKING CHANGE: web clients will need to update the type of the exposed script parameters that we changed.

+++ Library/src/Multiplayer/Script/ComponentBinding/AvatarSpaceComponentScriptInterface.cpp
@@ -33,14 +33,14 @@ DEFINE_SCRIPT_PROPERTY_STRING(AvatarSpaceComponent, AvatarId);
 DEFINE_SCRIPT_PROPERTY_STRING(AvatarSpaceComponent, UserId);
 DEFINE_SCRIPT_PROPERTY_STRING(AvatarSpaceComponent, AgoraUserId);
 DEFINE_SCRIPT_PROPERTY_STRING(AvatarSpaceComponent, CustomAvatarUrl);
-DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, csp::multiplayer::AvatarState, int64_t, State);
-DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, int64_t, int64_t, AvatarMeshIndex);
+DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, csp::multiplayer::AvatarState, int32_t, State);
+DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, int64_t, int32_t, AvatarMeshIndex);
 DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, bool, bool, IsHandIKEnabled);
 DEFINE_SCRIPT_PROPERTY_VEC3(AvatarSpaceComponent, TargetHandIKTargetLocation);
 DEFINE_SCRIPT_PROPERTY_VEC4(AvatarSpaceComponent, HandRotation);
 DEFINE_SCRIPT_PROPERTY_VEC4(AvatarSpaceComponent, HeadRotation);
 DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, float, float, WalkRunBlendPercentage);
 DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, float, float, TorsoTwistAlpha);
-DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, csp::multiplayer::AvatarPlayMode, int64_t, AvatarPlayMode);
+DEFINE_SCRIPT_PROPERTY_TYPE(AvatarSpaceComponent, csp::multiplayer::AvatarPlayMode, int32_t, AvatarPlayMode);
 
 } // namespace csp::multiplayer
